Criando o controlador e o service para mostrar o saldo disponivel(avaliable) e o saldo que a receber(waiting_funds)

+++ desafios/software-engineer-backend/psp-pagarme/src/models/transactions.model.ts
@@ -4,7 +4,7 @@ import { Table, Model, Column, DataType } from "sequelize-typescript";
 export class Transaction extends Model<Transaction> {
 
     @Column({
-        type: DataType.DECIMAL(20, 2),
+        type: DataType.FLOAT(20, 2),
         allowNull: false,
     })
     valor: number;
